Wrapped custom desmos queries to match the go desmos app fixed tests cleaned imports rebuild schemas put tests inside contract file split custo queries from query types
