+++ contracts/filter-posts/examples/schema.rs
@@ -1,10 +1,11 @@
-use std::env::current_dir;
-use std::fs::create_dir_all;
+use std::{env::current_dir, fs::create_dir_all};
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use filter_posts::msg::{HandleMsg, InitMsg, QueryMsg};
-use filter_posts::state::State;
+use filter_posts::{
+    msg::{HandleMsg, InitMsg, QueryMsg},
+    state::State,
+};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
