+++ packages/desmos/examples/schema.rs
@@ -1,10 +1,11 @@
-use std::env::{current_dir, set_current_dir};
-use std::fs::create_dir_all;
+use std::{env::current_dir, fs::create_dir_all};
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use desmos::custom_query::{PostsResponse, ReportsResponse};
-use desmos::types::{Attachment, OptionalData, PollAnswer, PollData, Post, Report};
+use desmos::{
+    query_types::{PostsResponse, ReportsResponse},
+    types::{Attachment, OptionalData, PollAnswer, PollData, Post, Report},
+};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
