+++ packages/desmos/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod custom_query;
+pub mod query_types;
 pub mod types;
