+++ packages/desmos/src/types.rs
@@ -1,7 +1,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-/// This file contains all the desmos related types used inside the contract
+/// This file contains all the desmos related types used inside desmos' contracts
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
